Stable Version of Zone Supervisor(DTO,SERVICE,MODEL,CONTROLLER,REPOSITORY). Issues of compatibility with recycleZone Class were addressed, and now I am testing GET and POST. Config package is commented and waiting for review due to error
