+++ src/main/java/com/example/proyectoFinal1/dto/ZoneSupervisorDTO.java
@@ -3,9 +3,5 @@ package com.example.proyectoFinal1.dto;
 import lombok.Data;
 @Data
 public class ZoneSupervisorDTO {
-    private Long ZoneSupervisorId;
-    private String ZoneSupervisorName;
-    private long ZoneSupervisorZoneId;
-
 }
