+++ src/main/java/com/example/proyectoFinal1/repository/ZoneSupervisorRepository.java
@@ -1,6 +1,5 @@
 package com.example.proyectoFinal1.repository;
 
-import com.example.proyectoFinal1.model.RecyclingZone;
 import com.example.proyectoFinal1.model.ZoneSupervisor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,5 +9,5 @@ import java.util.List;
 
 @Repository
 public interface ZoneSupervisorRepository extends JpaRepository<ZoneSupervisor, Long>{
-    List<RecyclingZone> zoneSupervisorRepository = new ArrayList<>();
+    List<ZoneSupervisor> zoneSupervisorRepository = new ArrayList<>();
 }
