+++ src/main/java/com/example/proyectoFinal1/service/ZoneSupervisorService.java
@@ -25,38 +25,29 @@ public class ZoneSupervisorService {
         zoneSupervisorRepository.deleteById(zoneSupervisorId);
     }
 
-    public List<ZoneSupervisorDTO> getAllZoneSupervisors(){
-        return zoneSupervisorRepository.findAll().
-                stream().map(this::convertEntityToDTO).collect(Collectors.toList());
-
-    }
+//    public List<ZoneSupervisorDTO> getAllZoneSupervisors(){
+//        return zoneSupervisorRepository.findAll().
+//                stream().map(this::convertEntityToDTO).collect(Collectors.toList());
+//
+//    }
 
     public void addNewZoneSupervisor(ZoneSupervisor zoneSupervisor) {
         zoneSupervisorRepository.save(zoneSupervisor);
 
     }
 
-    public ZoneSupervisorDTO getZoneSupervisorById(Long ZoneSupervisorId){
-        boolean exists = zoneSupervisorRepository.existsById(ZoneSupervisorId);
-        if (!exists) {
-            throw new IllegalStateException("RecyclingZone with id" + ZoneSupervisorId + "does not exist");
-        }
-        Optional<ZoneSupervisor> rz= zoneSupervisorRepository.findById(ZoneSupervisorId);
-        ZoneSupervisorDTO rzDTO = convertEntityToDTO(rz.get());
-        return rzDTO;
-    }
+//    public ZoneSupervisorDTO getZoneSupervisorById(Long ZoneSupervisorId){
+//        boolean exists = zoneSupervisorRepository.existsById(ZoneSupervisorId);
+//        if (!exists) {
+//            throw new IllegalStateException("RecyclingZone with id" + ZoneSupervisorId + "does not exist");
+//        }
+//        Optional<ZoneSupervisor> rz= zoneSupervisorRepository.findById(ZoneSupervisorId);
+//        ZoneSupervisorDTO rzDTO = convertEntityToDTO(rz.get());
+//        return rzDTO;
+//    }
 
     public ZoneSupervisor saveZonerSupervisor(ZoneSupervisor zoneSupervisor){
         return zoneSupervisorRepository.save(zoneSupervisor);
-    }
-    private ZoneSupervisorDTO convertEntityToDTO(ZoneSupervisor zoneSupervisor){
-        ZoneSupervisorDTO ZoneSupervisorDTO = new ZoneSupervisorDTO();
-        ZoneSupervisorDTO.setZoneSupervisorId(zoneSupervisor.getId());
-        ZoneSupervisorDTO.setZoneSupervisorName(zoneSupervisor.getName());
-        ZoneSupervisorDTO.setZoneSupervisorZoneId(zoneSupervisor.getZoneId());
-        //recyclingZoneDTO.setRecyclingZoneAdress(recyclingZone.getAdress());
-
-        return ZoneSupervisorDTO;
     }
